changed code based on review

+++ src/components/DeleteComment.jsx
@@ -1,24 +1,36 @@
 import React, { useState } from "react";
 import { deleteCommentByID } from "../api";
 
-export default function DeleteComment({ comment_id, setCommentExists }) {
+export default function DeleteComment({
+  comment_id,
+  setCommentExists,
+  author,
+  user,
+}) {
   const [loading, setloading] = useState(false);
   const [err, setErr] = useState(null);
+  const [allow, setAllow] = useState(true);
+
   const handleDelete = () => {
     setloading(true);
-    deleteCommentByID(comment_id).then((res) => {
-      if (res.status !== 204) {
-        setErr("Something Went Wrong Cant Delete Comment:(");
-      }
-      setCommentExists(false);
-      setloading(false);
-    });
+    setAllow(true);
+    if (user === author) {
+      deleteCommentByID(comment_id).then((res) => {
+        if (res.status !== 204) {
+          setErr("Something Went Wrong Cant Delete Comment:(");
+        }
+        setCommentExists(false);
+        setloading(false);
+      });
+    }
+    setAllow(false);
   };
   if (err) return <h4>{err}</h4>;
+  else if (!allow) return <h4>You can only delete your comments</h4>;
   else if (loading) return <h4>Deleting Comment</h4>;
   return (
     <div>
-      <button onClick={handleDelete}>DELETE COMMENT{comment_id}</button>
+      <button onClick={handleDelete}>DELETE COMMENT</button>
     </div>
   );
 }
